Making some simple fixes for better operation on 4808.

+++ src/nvm_flash.h
@@ -4,4 +4,4 @@
 #include <stdint.h>
 #include <avr/io.h>
 
-bool nvm_flash_write_page(uint16_t flash_addr, const uint8_t page[PROGMEM_PAGE_SIZE]);
+bool nvm_flash_write_page(uint32_t flash_addr, const uint8_t page[PROGMEM_PAGE_SIZE]);
